misc: move app to App folder and add polyfill types
