+++ .prettierrc.yaml
@@ -1,4 +1,4 @@
-semi: false
+semi: true
 tabWidth: 2
 trailingComma: 'all'
 singleQuote: true
