+++ package.json
@@ -31,7 +31,6 @@
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.3",
     "@rollup/plugin-replace": "^5.0.2",
-    "@types/chrome": "^0.0.224",
     "@types/firefox-webext-browser": "^111.0.0",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
@@ -50,6 +49,7 @@
     "@rollup/plugin-node-resolve": "^15.0.1",
     "@rollup/plugin-typescript": "^11.0.0",
     "@types/jest": "^29.5.0",
+    "@types/webextension-polyfill": "^0.10.0",
     "@typescript-eslint/eslint-plugin": "^5.55.0",
     "@typescript-eslint/parser": "^5.55.0",
     "autoprefixer": "^10.4.14",
