+++ src/pages/popup/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { createRoot } from 'react-dom/client'
-import App from './App'
+import App from '../../App/App'
 
 console.log('popup script')
